pyproject.toml
[tool.poetry]
name = "pyhton-sapling"
version = "0.1.0"
description = ""
authors = ["edmamerto"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = {extras = ["email"], version = "^2.6.3"}
structlog = "^24.1.0"
pytest-mock = "^3.12.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"

[tool.poetry.scripts]
main = "main:main"
test = "scripts.run_tests:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
pre-commit = "^3.6.2"
flake8 = "^7.0.0"
black = "^24.2.0"
isort = "^5.13.2"
mypy = "^1.8.0"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]

[tool.flake8]
max-line-length = 102
exclude = [
    ".git",
    "venv",
    "build",
    "dist",
    ".mypy_cache",
    ".tox",
    "_build",
    "buck-out",
]
ignore = ["E203", "W503"]

[tool.mypy]
files = ["src", "tests"]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
